requirements.txt
pandas
waitress
numpy
matplotlib
backtesting
bokeh
requests
flask
alpaca-trade-api
flask-restful
flasgger
alpaca-py
ta